requirements.txt
selenium>=3.141.0
pillow>=6.1.0
